refactor compact segment
