+++ src/query/storages/fuse/src/operations/mutation/compact/compact_aggregator.rs
