+++ src/query/storages/fuse/src/operations/mutation/compact/mod.rs
@@ -13,17 +13,17 @@
 // limitations under the License.
 
 mod block_compact_mutator;
+mod compact_aggregator;
 mod compact_meta;
 mod compact_part;
 mod compact_source;
 mod segment_compact_mutator;
-mod transform_compact_aggregator;
 
 pub use block_compact_mutator::BlockCompactMutator;
+pub use compact_aggregator::CompactAggregator;
 pub use compact_meta::CompactSourceMeta;
 pub use compact_part::CompactPartInfo;
 pub use compact_source::CompactSource;
 pub use segment_compact_mutator::SegmentCompactMutator;
 pub use segment_compact_mutator::SegmentCompactionState;
 pub use segment_compact_mutator::SegmentCompactor;
-pub use transform_compact_aggregator::CompactAggregator;
